Fix test `test_error_code_translation` on MacOS
* The test failed when sending in a `float` on MacOS aarch64.
* Sending a byte (`u8`) instead should work for all targets.

+++ libffi-rs/src/low.rs
@@ -772,7 +772,7 @@ mod test {
 
         // Provoke a FFI_BAD_ARGTYPE by bad values to `ffi_prep_cif_var`.
         let mut cif = ffi_cif::default();
-        let mut arg_types = [&raw mut types::float];
+        let mut arg_types = [&raw mut types::uint8];
 
         // SAFETY: Both `cif` and `args` are pointers to valid data.
         let result = unsafe {
